+++ runtime/vm/bootstrap_natives.h
@@ -66,9 +66,13 @@ namespace dart {
   V(SendPort_sendInternal_, 2)                                                 \
   V(Smi_bitNegate, 1)                                                          \
   V(Smi_bitLength, 1)                                                          \
+  V(Smi_trailingZeroBitCount, 1)                                               \
+  V(Smi_oneBitCount, 1)                                                    \
   V(SuspendState_instantiateClosureWithFutureTypeArgument, 2)                  \
   V(Mint_bitNegate, 1)                                                         \
   V(Mint_bitLength, 1)                                                         \
+  V(Mint_trailingZeroBitCount, 1)                                              \
+  V(Mint_oneBitCount, 1)                                                   \
   V(Developer_debugger, 2)                                                     \
   V(Developer_getIsolateIdFromSendPort, 1)                                     \
   V(Developer_getObjectId, 1)                                                  \